Move obsolete Send feature screens to appropriate package
