+++ app/src/main/java/com/tari/android/wallet/application/walletManager/WalletManager.kt
@@ -73,7 +73,7 @@ import com.tari.android.wallet.tor.TorConfig
 import com.tari.android.wallet.tor.TorProxyManager
 import com.tari.android.wallet.tor.TorProxyStateHandler
 import com.tari.android.wallet.ui.common.DialogManager
-import com.tari.android.wallet.ui.screen.send.finalize.FinalizeSendTxModel
+import com.tari.android.wallet.ui.screen.send.obsolete.finalize.FinalizeSendTxModel
 import com.tari.android.wallet.util.BroadcastEffectFlow
 import com.tari.android.wallet.util.Constants
 import com.tari.android.wallet.util.DebugConfig
