+++ app/src/main/java/com/tari/android/wallet/navigation/Navigation.kt
@@ -3,9 +3,9 @@ package com.tari.android.wallet.navigation
 import android.net.Uri
 import com.tari.android.wallet.data.contacts.Contact
 import com.tari.android.wallet.model.MicroTari
+import com.tari.android.wallet.model.TransactionData
 import com.tari.android.wallet.model.tx.Tx
 import com.tari.android.wallet.ui.screen.pinCode.PinCodeScreenBehavior
-import com.tari.android.wallet.ui.screen.send.common.TransactionData
 
 sealed class Navigation {
