--- app/src/main/java/com/tari/android/wallet/ui/screen/send/amountView/AmountStyle.kt
@@ -1,6 +1,0 @@
-package com.tari.android.wallet.ui.screen.send.amountView
-
-enum class AmountStyle {
-    Normal,
-    Warning
-}
