+++ app/src/main/java/com/tari/android/wallet/ui/screen/send/confirm/ConfirmViewModel.kt
@@ -3,10 +3,10 @@ package com.tari.android.wallet.ui.screen.send.confirm
 import androidx.lifecycle.SavedStateHandle
 import com.tari.android.wallet.R
 import com.tari.android.wallet.model.MicroTari
+import com.tari.android.wallet.model.TransactionData
 import com.tari.android.wallet.navigation.Navigation
 import com.tari.android.wallet.navigation.TariNavigator.Companion.PARAMETER_TRANSACTION
 import com.tari.android.wallet.ui.common.CommonViewModel
-import com.tari.android.wallet.ui.screen.send.common.TransactionData
 import com.tari.android.wallet.util.extension.getOrThrow
 import com.tari.android.wallet.util.shortString
 import kotlinx.coroutines.flow.MutableStateFlow
