+++ app/src/main/java/com/tari/android/wallet/ui/screen/send/obsolete/addNote/AddNoteViewModel.kt
@@ -1,10 +1,10 @@
-package com.tari.android.wallet.ui.screen.send.addNote
+package com.tari.android.wallet.ui.screen.send.obsolete.addNote
 
 import com.tari.android.wallet.data.network.NetworkConnectionStateHandler
 import com.tari.android.wallet.model.TariWalletAddress
+import com.tari.android.wallet.model.TransactionData
 import com.tari.android.wallet.navigation.Navigation
 import com.tari.android.wallet.ui.common.CommonViewModel
-import com.tari.android.wallet.ui.screen.send.common.TransactionData
 import javax.inject.Inject
 
 class AddNoteViewModel : CommonViewModel() {
