+++ app/src/main/java/com/tari/android/wallet/ui/screen/send/obsolete/amountView/AmountStyle.kt
@@ -1,0 +1,6 @@
+package com.tari.android.wallet.ui.screen.send.obsolete.amountView
+
+enum class AmountStyle {
+    Normal,
+    Warning
+}
