+++ app/src/main/java/com/tari/android/wallet/ui/screen/send/obsolete/requestTari/RequestTariFragment.kt
@@ -1,4 +1,4 @@
-package com.tari.android.wallet.ui.screen.send.requestTari
+package com.tari.android.wallet.ui.screen.send.obsolete.requestTari
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -15,7 +15,7 @@ import com.tari.android.wallet.ui.dialog.modular.ModularDialogArgs
 import com.tari.android.wallet.ui.dialog.modular.modules.button.ButtonModule
 import com.tari.android.wallet.ui.dialog.modular.modules.button.ButtonStyle
 import com.tari.android.wallet.ui.dialog.modular.modules.shareQr.ShareQrCodeModule
-import com.tari.android.wallet.ui.screen.send.addAmount.keyboard.KeyboardController
+import com.tari.android.wallet.ui.screen.send.obsolete.addAmount.keyboard.KeyboardController
 import com.tari.android.wallet.util.extension.hideKeyboard
 import com.tari.android.wallet.util.extension.setOnThrottledClickListener
 import com.tari.android.wallet.util.extension.setVisible
