+++ app/src/main/res/layout/fragment_add_amount.xml
@@ -103,7 +103,7 @@
                     android:paddingBottom="@dimen/add_amount_balance_warning_container_bottom_padding"
                     tools:ignore="UselessParent">
 
-                    <com.tari.android.wallet.ui.screen.send.amountView.AmountView
+                    <com.tari.android.wallet.ui.screen.send.obsolete.amountView.AmountView
                         android:id="@+id/available_balance_container_view"
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
